Conflicts:
	CCTM/src/cio/centralized_io_module.F
	CCTM/src/emis/emis/EMIS_DEFN.F
	CCTM/src/gas/ebi_cb6r3_ae6_aq/degrade_data.F
	CCTM/src/gas/ebi_cb6r3_ae7_aq/degrade_data.F
	CCTM/src/gas/ebi_cb6r3m_ae7_kmtbr/degrade_data.F
	CCTM/src/gas/smvgear/degrade_data.F
	UTIL/create_ebi/degrade_codes_serial-RXNST/degrade_data.F
	bldit_project.csh

+++ CCTM/scripts/bldit_cctm.csh
@@ -39,11 +39,6 @@
  set Mechs   = $CCTM_SRC/MECHS         #> location of the chemistry mechanism include files
  setenv REPOROOT $CCTM_SRC
 
-#> Working directory and Version IDs
- set VRSN  = v532                      #> model configuration ID
- set EXEC  = CCTM_${VRSN}.exe          #> executable name
- set CFG   = CCTM_${VRSN}.cfg          #> configuration file name
-
 #> Controls for managing the source code and MPI compilation
 set CompileBLDMAKE                     #> Recompile the BLDMAKE utility from source
                                        #>   comment out to use an existing BLDMAKE executable
@@ -78,6 +73,19 @@ set make_options = "-j"                #> additional options for make command if
 
 #> Potential vorticity free-troposphere O3 scaling
 #set potvortO3
+
+#> Working directory and Version IDs
+ if ( $?ISAM_CCTM ) then
+     set VRSN  = v532_ISAM             #> model configuration ID for CMAQ_ISAM
+ else if ( $?DDM3D_CCTM ) then
+     set VRSN = v532_DDM3D             #> model configuration ID for CMAQ_DDM
+ else
+     set VRSN = v532                   #> model configuration ID for CMAQ
+ endif
+ 
+ set EXEC  = CCTM_${VRSN}.exe          #> executable name
+ set CFG   = CCTM_${VRSN}.cfg          #> configuration file name
+
 
 #========================================================================
 #> CCTM Science Modules
@@ -687,13 +695,19 @@ set Cfile = ${Bld}/${CFG}.bld      # Config Filename
     set bld_flags = "${bld_flags} -isam_cctm"
  endif
 
+ if ( $?build_twoway ) then
+   set bld_flags = "${bld_flags} -twoway"
+ endif
+
 #> Run BLDMAKE with source code in build directory
  $Blder $bld_flags $Cfile   
 
 #> Rename Makefile to specify compiler option and link back to Makefile
- mv Makefile Makefile.$compilerString
- if ( -e Makefile.$compilerString && -e Makefile ) rm Makefile
- ln -s Makefile.$compilerString Makefile
+ if ( ! $?build_twoway ) then
+    mv Makefile Makefile.$compilerString
+    if ( -e Makefile.$compilerString && -e Makefile ) rm Makefile
+    ln -s Makefile.$compilerString Makefile
+ endif
 
 #> Alert user of error in BLDMAKE if it ocurred
  if ( $status != 0 ) then
